Madara: Add related manga support to Madara and clean up related manga methods in other sources
