+++ lib-multisrc/madara/build.gradle.kts
@@ -1,3 +1,5 @@
+kmkBaseVersionCode = 1
+
 plugins {
     id("lib-multisrc")
 }
