+++ src/ko/manhwaraw/src/eu/kanade/tachiyomi/extension/ko/manhwaraw/ManhwaRaw.kt
@@ -220,23 +220,4 @@ class ManhwaRaw : Madara("ManhwaRaw", "https://manhwaraw.com", "ko") {
 
         return POST("$baseUrl/wp-admin/admin-ajax.php", xhrHeaders, form)
     }
-
-    override fun relatedMangaListSelector() = ".related-manga .related-reading-wrap"
-
-    override fun relatedMangaFromElement(element: Element): SManga {
-        val manga = SManga.create()
-
-        with(element) {
-            selectFirst(".widget-title a")!!.let {
-                manga.setUrlWithoutDomain(it.attr("abs:href"))
-                manga.title = it.ownText()
-            }
-
-            selectFirst("img")?.let {
-                manga.thumbnail_url = imageFromElement(it)
-            }
-        }
-
-        return manga
-    }
 }
